Fix propagate constants rule.
