+++ arangod/Aql/CMakeLists.txt
@@ -106,6 +106,7 @@ add_library(arango_aql STATIC
   NonConstExpression.cpp
   NonConstExpressionContainer.cpp
   Optimizer.cpp
+  OptimizerRulePropagateConstantAttributes.cpp
   OptimizerRuleReplaceEqualAttributeAccess.cpp
   OptimizerRules.cpp
   OptimizerRulesCluster.cpp
